fix(ui): persist new sessions eagerly and refresh titles on re-entry
- Move sessions.create from after the first prompt resolves to right after
  newSession returns. A refresh or navigation mid-turn no longer leaves
  the session missing from the sidebar (#21).
- Render "(no title)" in muted style when the harness hasn't named the
  session yet, instead of falling back to a session-id slice that read
  like a title.
- Add refetchOnMount: "always" to the sessions list query so returning
  to a chat view picks up titles set after the first turn without a
  manual refresh.


fix(ui): disambiguate untitled sessions and soften save-failure toast

- Append a short id suffix to the placeholder so multiple untitled rows
  remain distinguishable: "(no title · abcd1234)".
- Lift the title span's className ternary into a titleClass/titleLabel
  pair above the JSX for readability.
- Reword the session-create failure toast: the upsert now runs mid-turn,
  so the previous "won't appear in the list" wording read alarmingly
  while the agent was still streaming.


refactor(ui): extract SessionRow into its own file

Apply the touch-it-migrate-it rule to the bits of sessions-sidebar.tsx
touched in this PR. Move SessionRow into a sibling file with an
explicit `interface Props` typed on SessionView. Drop the now-unused
SessionMode/SessionType/Trash2/useEffect/useRef/useState imports from
the parent file.

The broader SessionsSidebar cleanup (inline useStore selectors, JSX
weight) is project-wide and belongs in a separate PR.
