+++ packages/ui/src/modules/sessions/hooks/use-acp-session-engagement.ts
@@ -1,9 +1,13 @@
 import type { ClientSideConnection } from "@agentclientprotocol/sdk/dist/acp.js";
 import type { McpServer } from "@agentclientprotocol/sdk/dist/schema/types.gen.js";
+import { SessionMode } from "api-server-api";
 import { useCallback, useRef } from "react";
 
+import { api } from "../../../api.js";
+import { queryClient } from "../../../query-client.js";
 import { useStore } from "../../../store.js";
 import type { SessionConfigPayload } from "../../acp/types.js";
+import { acpSessionsKeys } from "../api/queries.js";
 
 /**
  * Owns the "engage a live ACP connection with the active session" decision.
@@ -37,6 +41,7 @@ export function useAcpSessionEngagement(
 } {
   const setSessionId = useStore((s) => s.setSessionId);
   const addLog = useStore((s) => s.addLog);
+  const showToast = useStore((s) => s.showToast);
 
   const engagedSessionIdRef = useRef<string | null>(null);
 
@@ -63,9 +68,23 @@ export function useAcpSessionEngagement(
       setSessionId(s.sessionId);
       engagedSessionIdRef.current = s.sessionId;
       addLog("session", { sessionId: s.sessionId });
+      // Persist to the platform DB as soon as the runtime returns a session
+      // id — refreshing or navigating away mid-turn must still leave the
+      // session in the sidebar. Fire-and-forget; engage only ever runs from
+      // a user-initiated send, so this can't create rows for chats the user
+      // never sent in.
+      api.sessions.create
+        .mutate({ sessionId: s.sessionId, instanceId: selectedInstance, mode: SessionMode.Chat })
+        .then(() => queryClient.invalidateQueries({ queryKey: acpSessionsKeys.all }))
+        .catch((err) => {
+          showToast({
+            kind: "warning",
+            message: `Couldn't save this session to your history: ${err instanceof Error ? err.message : "sync failed"}`,
+          });
+        });
       await applySavedPreferences(conn, s.sessionId, s);
     }
-  }, [selectedInstance, selectedMcpServers, captureSessionConfig, applySavedPreferences, setSessionId, addLog]);
+  }, [selectedInstance, selectedMcpServers, captureSessionConfig, applySavedPreferences, setSessionId, addLog, showToast]);
 
   const clear = useCallback(() => {
     engagedSessionIdRef.current = null;
